+++ lib/stream.h
@@ -78,7 +78,7 @@ int pstream_open_with_default_port(const char *name,
                                    struct pstream **,
                                    uint8_t dscp);
 bool stream_parse_target_with_default_port(const char *target,
-                                           uint16_t default_port,
+                                           int default_port,
                                            struct sockaddr_storage *ss);
 int stream_or_pstream_needs_probes(const char *name);
